fix: Dapper constructor type mismatches crashing SSO login
user_statistics.OnlineTime is a MySQL int, but HabboStats' constructor
declared its onlineTime parameter as double - Dapper's Query<T> maps
straight onto the constructor and requires an exact type match per
column, so login crashed on every SSO ticket with "A parameterless
default constructor or one matching signature ... is required for
Plus.HabboHotel.Users.HabboStats materialization" (PlusEnvironment
never retries this - it just fails the login).

Fixed OnlineTime's type (double -> int) but hit the same problem one
column over: quest_id is `int unsigned`, which fails to match the
`int questId` parameter the same way. Rather than keep chasing
column-by-column type matches (and rather than change HabboStats.QuestId
to uint, which would ripple `int` casts through every call site in
QuestManager and its packet composers), reworked
HabboStatsService.LoadHabboStats to query as `dynamic` and build the
HabboStats explicitly with Convert.ToInt32 per field - sidesteps
Dapper's strict constructor-materialization entirely so it's no longer
sensitive to the exact signed/unsigned width MySqlConnector reports
for a given column.

Confirmed via the emulator's debug packet log: before this fix, every
SsoTicketEvent for a real login threw and the client never got past
handshake; after, AuthenticationOkComposer and the full post-login
packet burst send cleanly.

+++ HabboHotel/Users/HabboStats.cs
@@ -2,7 +2,7 @@
 
 public class HabboStats
 {
-    public HabboStats(int roomVisits, double onlineTime, int respect, int respectGiven, int giftsGiven, int giftsReceived, int dailyRespectPoints, int dailyPetRespectPoints, int achievementPoints,
+    public HabboStats(int roomVisits, int onlineTime, int respect, int respectGiven, int giftsGiven, int giftsReceived, int dailyRespectPoints, int dailyPetRespectPoints, int achievementPoints,
         int questId, int questProgress, int groupId, string respectsTimestamp, int forumPosts)
     {
         RoomVisits = roomVisits;
@@ -22,7 +22,7 @@ public class HabboStats
     }
 
     public int RoomVisits { get; set; }
-    public double OnlineTime { get; set; }
+    public int OnlineTime { get; set; }
     public int Respect { get; set; }
     public int RespectGiven { get; set; }
     public int GiftsGiven { get; set; }
